chore(exception): Cleanup API exception class
