+++ tox.ini
@@ -5,7 +5,7 @@ envlist = lint, py27, py34, py35, py36, py37
 [testenv:lint]
 basepython = python2.7
 deps = pylint
-commands = pylint ibm_cloud_sdk_core test examples
+commands = pylint ibm_cloud_sdk_core test
 
 [testenv]
 passenv = TOXENV CI TRAVIS*
